requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
bcrypt==4.0.1
eventlet==0.33.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0   # optional, for loading .env file locally